+++ src/models/state.model.ts
@@ -6,5 +6,4 @@ export interface IExtensionState {
   context?: ExtensionContext;
   environment?: Environment;
   commons?: Commons;
-  instanceID: string;
 }
