+++ src/service/github.service.ts
@@ -3,9 +3,9 @@
 import * as GitHubApi from "@octokit/rest";
 import * as HttpsProxyAgent from "https-proxy-agent";
 import * as vscode from "vscode";
-import Commons from "../commons";
 import { state } from "../state";
 import { File } from "./file.service";
+import { LoggerService } from "./logger.service";
 
 interface IEnv {
   [key: string]: string | undefined;
@@ -133,9 +133,9 @@ export class GitHubService {
     const promise = this.github.gists.get({ gist_id: GIST });
     const res = await promise.catch(err => {
       if (String(err).includes("HttpError: Not Found")) {
-        return Commons.LogException(err, "Sync: Invalid Gist ID", true);
+        return LoggerService.LogException(err, "Sync: Invalid Gist ID", true);
       }
-      Commons.LogException(err, state.commons.ERROR_MESSAGE, true);
+      LoggerService.LogException(err, state.commons.ERROR_MESSAGE, true);
     });
     if (res) {
       return res;
@@ -185,9 +185,9 @@ export class GitHubService {
 
     const res = await promise.catch(err => {
       if (String(err).includes("HttpError: Not Found")) {
-        return Commons.LogException(err, "Sync: Invalid Gist ID", true);
+        return LoggerService.LogException(err, "Sync: Invalid Gist ID", true);
       }
-      Commons.LogException(err, state.commons.ERROR_MESSAGE, true);
+      LoggerService.LogException(err, state.commons.ERROR_MESSAGE, true);
     });
 
     if (res) {
