+++ src/service/settings.service.ts
@@ -1,0 +1,393 @@
+import * as vscode from "vscode";
+import localize from "../localize";
+import { CustomConfig } from "../models/customConfig.model";
+import { ExtensionConfig } from "../models/extensionConfig.model";
+import { LocalConfig } from "../models/localConfig.model";
+import { state } from "../state";
+import { File, FileService } from "./file.service";
+import { GitHubService } from "./github.service";
+import { LoggerService } from "./logger.service";
+
+export class SettingsService {
+  public async ResetSettings(): Promise<void> {
+    const extSettings = new ExtensionConfig();
+    const localSettings = new CustomConfig();
+
+    vscode.window.setStatusBarMessage(
+      localize("cmd.resetSettings.info.resetting"),
+      2000
+    );
+
+    try {
+      const extSaved: boolean = await state.commons.SaveSettings(extSettings);
+      const customSaved: boolean = await state.commons.SetCustomSettings(
+        localSettings
+      );
+
+      if (extSaved && customSaved) {
+        vscode.window.showInformationMessage(
+          localize("cmd.resetSettings.info.settingClear")
+        );
+      }
+    } catch (err) {
+      LoggerService.LogException(
+        err,
+        "Sync: Unable to clear settings. Error logged to the console.",
+        true
+      );
+    }
+  }
+
+  public async OpenAdvancedOptions() {
+    const setting: ExtensionConfig = await state.commons.GetSettings();
+    const customSettings: CustomConfig = await state.commons.GetCustomSettings();
+    if (customSettings == null) {
+      vscode.window
+        .showInformationMessage(
+          localize("cmd.otherOptions.triggerReset"),
+          localize("common.button.yes")
+        )
+        .then(val => {
+          if (val === localize("common.button.yes")) {
+            vscode.commands.executeCommand("extension.resetSettings");
+          }
+        });
+    }
+    const localSetting = new LocalConfig();
+    const tokenAvailable: boolean =
+      customSettings.token != null && customSettings.token !== "";
+    const gistAvailable: boolean = setting.gist != null && setting.gist !== "";
+
+    const items: string[] = [
+      "cmd.otherOptions.editLocalSetting",
+      "cmd.otherOptions.shareSetting",
+      "cmd.otherOptions.downloadSetting",
+      "cmd.otherOptions.toggleForceDownload",
+      "cmd.otherOptions.toggleForceUpload",
+      "cmd.otherOptions.toggleAutoUpload",
+      "cmd.otherOptions.toggleAutoDownload",
+      "cmd.otherOptions.toggleSummaryPage",
+      "cmd.otherOptions.customizedSync",
+      "cmd.otherOptions.downloadCustomFile",
+      "cmd.otherOptions.joinCommunity",
+      "cmd.otherOptions.openIssue",
+      "cmd.otherOptions.releaseNotes"
+    ].map(localize);
+
+    let selectedItem: number = 0;
+    let settingChanged: boolean = false;
+
+    const item = await vscode.window.showQuickPick(items);
+
+    // if not pick anyone, do nothing
+    if (!item) {
+      return;
+    }
+
+    const index = items.findIndex(v => v === item);
+
+    const handlerMap = {
+      0: async () => {
+        const file: vscode.Uri = vscode.Uri.file(
+          state.environment.FILE_CUSTOMIZEDSETTINGS
+        );
+        const document = await vscode.workspace.openTextDocument(file);
+        await vscode.window.showTextDocument(
+          document,
+          vscode.ViewColumn.One,
+          true
+        );
+      },
+      1: async () => {
+        // share public gist
+        const answer = await vscode.window.showInformationMessage(
+          localize("cmd.otherOptions.shareSetting.beforeConfirm"),
+          "Yes"
+        );
+
+        if (answer === "Yes") {
+          localSetting.publicGist = true;
+          settingChanged = true;
+          setting.gist = "";
+          selectedItem = 1;
+          customSettings.downloadPublicGist = false;
+          await state.commons.SetCustomSettings(customSettings);
+        }
+      },
+      2: async () => {
+        // Download Settings from Public GIST
+        selectedItem = 2;
+        customSettings.downloadPublicGist = true;
+        settingChanged = true;
+        await state.commons.SetCustomSettings(customSettings);
+      },
+      3: async () => {
+        // toggle force download
+        selectedItem = 3;
+        settingChanged = true;
+        setting.forceDownload = !setting.forceDownload;
+      },
+      4: async () => {
+        // toggle force upload
+        selectedItem = 4;
+        settingChanged = true;
+        setting.forceUpload = !setting.forceUpload;
+      },
+      5: async () => {
+        // toggle auto upload
+        selectedItem = 5;
+        settingChanged = true;
+        setting.autoUpload = !setting.autoUpload;
+      },
+      6: async () => {
+        // auto download on startup
+        selectedItem = 6;
+        settingChanged = true;
+        if (!setting) {
+          vscode.commands.executeCommand("extension.HowSettings");
+          return;
+        }
+        if (!gistAvailable) {
+          vscode.commands.executeCommand("extension.HowSettings");
+          return;
+        }
+
+        setting.autoDownload = !setting.autoDownload;
+      },
+      7: async () => {
+        // page summary toggle
+        selectedItem = 7;
+        settingChanged = true;
+
+        if (!tokenAvailable || !gistAvailable) {
+          vscode.commands.executeCommand("extension.HowSettings");
+          return;
+        }
+        setting.quietSync = !setting.quietSync;
+      },
+      8: async () => {
+        // add customized sync file
+        const options: vscode.InputBoxOptions = {
+          ignoreFocusOut: true,
+          placeHolder: localize("cmd.otherOptions.customizedSync.placeholder"),
+          prompt: localize("cmd.otherOptions.customizedSync.prompt")
+        };
+        const input = await vscode.window.showInputBox(options);
+
+        if (input) {
+          const fileName: string = FileService.ExtractFileName(input);
+          if (fileName === "") {
+            return;
+          }
+          customSettings.customFiles[fileName] = input;
+          const done: boolean = await state.commons.SetCustomSettings(
+            customSettings
+          );
+          if (done) {
+            vscode.window.showInformationMessage(
+              localize("cmd.otherOptions.customizedSync.done", fileName)
+            );
+          }
+        }
+      },
+      9: async () => {
+        // Import customized sync file to workspace
+        const customFiles = await this.CustomFilesFromGist(
+          customSettings,
+          setting
+        );
+        if (customFiles.length < 1) {
+          return;
+        }
+        const options: vscode.QuickPickOptions = {
+          ignoreFocusOut: true,
+          placeHolder: localize(
+            "cmd.otherOptions.downloadCustomFile.placeholder"
+          )
+        };
+        const fileName = await vscode.window.showQuickPick(
+          customFiles.map(file => {
+            return file.fileName;
+          }),
+          options
+        );
+        // if not pick anyone, do nothing
+        if (!fileName) {
+          return;
+        }
+        const selected = customFiles.find(f => {
+          return f.fileName === fileName;
+        });
+        if (selected && vscode.workspace.rootPath) {
+          const downloadPath = FileService.ConcatPath(
+            vscode.workspace.rootPath,
+            selected.fileName
+          );
+          const done = await FileService.WriteFile(
+            downloadPath,
+            selected.content
+          );
+          if (done) {
+            vscode.window.showInformationMessage(
+              localize("cmd.otherOptions.downloadCustomFile.done", downloadPath)
+            );
+          }
+        }
+      },
+      10: async () => {
+        vscode.commands.executeCommand(
+          "vscode.open",
+          vscode.Uri.parse(
+            "https://join.slack.com/t/codesettingssync/shared_invite/enQtMzE3MjY5NTczNDMwLTYwMTIwNGExOGE2MTJkZWU0OTU5MmI3ZTc4N2JkZjhjMzY1OTk5OGExZjkwMDMzMDU4ZTBlYjk5MGQwZmMyNzk"
+          )
+        );
+      },
+      11: async () => {
+        vscode.commands.executeCommand(
+          "vscode.open",
+          vscode.Uri.parse(
+            "https://github.com/shanalikhan/code-settings-sync/issues/new"
+          )
+        );
+      },
+      12: async () => {
+        vscode.commands.executeCommand(
+          "vscode.open",
+          vscode.Uri.parse(
+            "http://shanalikhan.github.io/2016/05/14/Visual-studio-code-sync-settings-release-notes.html"
+          )
+        );
+      }
+    };
+
+    try {
+      await handlerMap[index]();
+      if (settingChanged) {
+        if (selectedItem === 1) {
+          await state.commons.HandleStopWatching();
+        }
+        await state.commons
+          .SaveSettings(setting)
+          .then((added: boolean) => {
+            if (added) {
+              const callbackMap = {
+                1: async () => {
+                  return await vscode.commands.executeCommand(
+                    "extension.updateSettings",
+                    "publicGIST"
+                  );
+                },
+                2: async () => {
+                  return await vscode.window.showInformationMessage(
+                    localize("cmd.otherOptions.warning.tokenNotRequire")
+                  );
+                },
+                3: async () => {
+                  const message = setting.forceDownload
+                    ? "cmd.otherOptions.toggleForceDownload.on"
+                    : "cmd.otherOptions.toggleForceDownload.off";
+                  return vscode.window.showInformationMessage(
+                    localize(message)
+                  );
+                },
+                4: async () => {
+                  const message = setting.forceUpload
+                    ? "cmd.otherOptions.toggleForceUpload.on"
+                    : "cmd.otherOptions.toggleForceUpload.off";
+                  return vscode.window.showInformationMessage(
+                    localize(message)
+                  );
+                },
+                5: async () => {
+                  const message = setting.autoUpload
+                    ? "cmd.otherOptions.toggleAutoUpload.on"
+                    : "cmd.otherOptions.toggleAutoUpload.off";
+                  return vscode.window.showInformationMessage(
+                    localize(message)
+                  );
+                },
+                6: async () => {
+                  const message = setting.autoDownload
+                    ? "cmd.otherOptions.toggleAutoDownload.on"
+                    : "cmd.otherOptions.toggleAutoDownload.off";
+                  return vscode.window.showInformationMessage(
+                    localize(message)
+                  );
+                },
+                7: async () => {
+                  const message = setting.quietSync
+                    ? "cmd.otherOptions.quietSync.on"
+                    : "cmd.otherOptions.quietSync.off";
+                  return vscode.window.showInformationMessage(
+                    localize(message)
+                  );
+                }
+              };
+
+              if (callbackMap[selectedItem]) {
+                return callbackMap[selectedItem]();
+              }
+            } else {
+              return vscode.window.showErrorMessage(
+                localize("cmd.otherOptions.error.toggleFail")
+              );
+            }
+          })
+          .catch(err => {
+            LoggerService.LogException(
+              err,
+              "Sync : Unable to toggle. Please open an issue.",
+              true
+            );
+          });
+      }
+    } catch (err) {
+      LoggerService.LogException(err, "Error", true);
+      return;
+    }
+  }
+
+  public async OpenHelp() {
+    return vscode.env.openExternal(
+      vscode.Uri.parse(
+        "https://github.com/shanalikhan/code-settings-sync/wiki/Settings-Guide"
+      )
+    );
+  }
+
+  private async CustomFilesFromGist(
+    customSettings: CustomConfig,
+    syncSetting: ExtensionConfig
+  ): Promise<File[]> {
+    const github = new GitHubService(
+      customSettings.token,
+      customSettings.githubEnterpriseUrl
+    );
+    const res = await github.ReadGist(syncSetting.gist);
+    if (!res) {
+      return [];
+    }
+    const keys = Object.keys(res.data.files);
+    const customFiles: File[] = [];
+    keys.forEach(gistName => {
+      if (res.data.files[gistName]) {
+        if (res.data.files[gistName].content) {
+          const prefix = FileService.CUSTOMIZED_SYNC_PREFIX;
+          if (gistName.indexOf(prefix) > -1) {
+            const fileName = gistName.split(prefix).join(""); // |customized_sync|.htmlhintrc => .htmlhintrc
+            const f: File = new File(
+              fileName,
+              res.data.files[gistName].content,
+              fileName in customSettings.customFiles
+                ? customSettings.customFiles[fileName]
+                : null,
+              gistName
+            );
+            customFiles.push(f);
+          }
+        }
+      }
+    });
+    return customFiles;
+  }
+}
