+++ src/state.ts
@@ -1,5 +1,3 @@
 import { IExtensionState } from "./models/state.model";
 
-export const state: IExtensionState = {
-  instanceID: Math.random().toString()
-};
+export const state: IExtensionState = {};
